IPC: IPC4: Add File and dcblock to UUID map
This addition allows to load file components for DAI
and host copier replacement for testbench. DC blocker
is added as process component to allow testing of this
component as example. Support for other processing
components will need a new solution to pass the UUIDs
over IPC messages instead of the 8-bit index value.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>