Tools: Testbench: Add support for parsing controls from topology
This adds find or byte controls and set to components after
initialize. Also other control types mixer, switch, enum are
parsed but not yet applied to components.

The code is copied and adapted from SOF plugin. As addition the
bytes control is split to several messages if the IPC message
size limit is reached.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>