Tools: Testbench: Add large byte control messages split
If message is larger than TB_IPC4_MAX_MSG_SIZE split it to
multiple parts.

For the byte control split messages to work the testbench
TB_IPC4_MAX_MSG_SIZE need to be set to MAILBOX_DSPBOX_SIZE
that is the assumed size of a fragment size elsewhere.

Since in testbench (and SOF plugin) the message size is consumed
by both header and payload the chunk size is smaller than in real
device. The function module_set_large_config() assumes that
fragment size for middle and first is MAILBOX_DSPBOX_SIZE. With
testbench the IPC4 header size needs to be subtracted from it.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>