Tools: Testbench: Switch testbench to IPC4 mode
The library_defconfig is changed to CONFIG_IPC_MAJOR_4=y.

To prepare release of testbench with SOF the executable name
is changed to sof-testbench4 for IPC4. The name is selected
based on choice in library_defconfig.

The script comp_run.sh for process_test.m and other test scripts
is changed to execute the IPC4 testbench version with use of
other topology v2 topologies sof-hda-benchmark-<comp><bits>.tplg.

the script host-testbench.sh is changed to run tests for
component names in IPC4 test topologies.

The github workflow is changed due to change in used topologies
for the test. The scripts/build-tools.sh need to be executed
with "-Y" option to build topology2 files. Also need to use
alsatplg from docker. The octave-io package is not needed for
this step.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>