Audio: Aria: Improvements to configuration blob handler
It depends on architecture but size_t and uint32_t are not
necessarily bits compatible (on 64bit), so memcpy is replaced
with type casts for the attenuation parameter copy.

A similar check as in aria_init() is added to avoid an illegal
value to break the processing algorithm.

An info level trace is added to see in trace if the blob
is applied by the Linux topology or in run-time.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>