Audio: SRC: Add HiFi5 version, as copy of HiFi4 version
Copying the file as such from HiFi4 and adding it to build
as preparation for optimizations.

The src_config.h is updated to follow the kconfig of FILTER
(CONFIG_FILTER_HIFI_NONE, CONFIG_FILTER_HIFI_3, etc.). The
gcc build is no more forced to use 16 bit SRC mode. It's a
limitation from HiFi2 platform BYT. The new platforms are
powerful enough to run it.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>