ipc4: pipeline: Add direction fields in the pipeline IPC msg
Add a couple of fields in the pipeline IPC message extension to indicate
if the pipeline direction is set and the direction. This is needed for
hostless pipelines such as the following:

tone generator -> mixin -> mixout -> gain -> DAI copier

In the above case, since there is no host copier at the source, the
component direction cannot be propagated.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>